start splitting up requestData into requestData and responseDetails
